reworked the loadBru stuff and tried to slim down main.go. Reworked a bit of stuff, going to try and tabs next.
