+++ src/main.go
@@ -168,284 +168,16 @@ func initialModel() model {
 }
 
 func (m *model) loadBruFiles() {
-	m.collections = []panels.CollectionItem{}
-	m.bruRequests = []*panels.BruRequest{}
-
 	collectionsDir, err := getCollectionsDir()
 	if err != nil {
+		m.collections = []panels.CollectionItem{}
+		m.bruRequests = []*panels.BruRequest{}
 		return
 	}
 
-	// Check if collections directory exists and has content
-	if _, err := os.Stat(collectionsDir); os.IsNotExist(err) {
-		return
-	}
-
-	// First, collect all directories (collections)
-	dirEntries, err := os.ReadDir(collectionsDir)
-	if err != nil {
-		return
-	}
-
-	// Group requests by collection and then by tag
-	collections := make(map[string]map[string][]*panels.BruRequest)
-	requestPaths := make(map[*panels.BruRequest]string)
-
-	// Add collection folders first and load their requests
-	for _, entry := range dirEntries {
-		if entry.IsDir() {
-			collectionName := entry.Name()
-			collectionPath := filepath.Join(collectionsDir, collectionName)
-			
-			collections[collectionName] = make(map[string][]*panels.BruRequest)
-
-			// Load .bru files from this collection
-			collectionFiles, err := os.ReadDir(collectionPath)
-			if err != nil {
-				continue
-			}
-
-			for _, file := range collectionFiles {
-				if strings.HasSuffix(file.Name(), ".bru") {
-					bruPath := filepath.Join(collectionPath, file.Name())
-					
-					bruFile, err := os.Open(bruPath)
-					if err != nil {
-						continue
-					}
-
-					parser := NewBruParser(bruFile)
-					request, err := parser.Parse()
-					bruFile.Close()
-					if err != nil {
-						continue
-					}
-
-					m.bruRequests = append(m.bruRequests, request)
-					requestPaths[request] = bruPath
-
-					// Group by tags, or use "untagged" if no tags
-					if len(request.Tags) == 0 {
-						collections[collectionName]["untagged"] = append(collections[collectionName]["untagged"], request)
-					} else {
-						for _, tag := range request.Tags {
-							collections[collectionName][tag] = append(collections[collectionName][tag], request)
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Add any standalone .bru files in the root collections directory
-	rootRequests := make(map[string][]*panels.BruRequest)
-	for _, entry := range dirEntries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".bru") {
-			bruPath := filepath.Join(collectionsDir, entry.Name())
-			
-			bruFile, err := os.Open(bruPath)
-			if err != nil {
-				continue
-			}
-
-			parser := NewBruParser(bruFile)
-			request, err := parser.Parse()
-			bruFile.Close()
-			if err != nil {
-				continue
-			}
-
-			m.bruRequests = append(m.bruRequests, request)
-			requestPaths[request] = bruPath
-
-			// Group by tags, or use "untagged" if no tags
-			if len(request.Tags) == 0 {
-				rootRequests["untagged"] = append(rootRequests["untagged"], request)
-			} else {
-				for _, tag := range request.Tags {
-					rootRequests[tag] = append(rootRequests[tag], request)
-				}
-			}
-		}
-	}
-
-	// Build the display list with tag grouping
-	requestIndexMap := make(map[*panels.BruRequest]int)
-	for i, req := range m.bruRequests {
-		requestIndexMap[req] = i
-	}
-
-	// Add collection folders with tag grouping
-	collectionNames := make([]string, 0, len(collections))
-	for name := range collections {
-		collectionNames = append(collectionNames, name)
-	}
-	sort.Strings(collectionNames)
-
-	for _, collectionName := range collectionNames {
-		tags := collections[collectionName]
-		
-		// Only add collection folder if it has requests
-		hasRequests := false
-		for _, requests := range tags {
-			if len(requests) > 0 {
-				hasRequests = true
-				break
-			}
-		}
-		
-		if !hasRequests {
-			continue
-		}
-
-		// Add collection folder header
-		m.collections = append(m.collections, panels.CollectionItem{
-			Name:         "📁 " + collectionName,
-			Type:         "folder",
-			FilePath:     filepath.Join(collectionsDir, collectionName),
-			IsFolder:     true,
-			IsTagGroup:   false,
-			RequestIndex: -1,
-			IsExpanded:   false, // Collapsed by default
-			IsVisible:    true,  // Folders are always visible
-		})
-
-		// Get sorted tag names
-		tagNames := make([]string, 0, len(tags))
-		for tagName := range tags {
-			if len(tags[tagName]) > 0 { // Only include tags that have requests
-				tagNames = append(tagNames, tagName)
-			}
-		}
-		sort.Strings(tagNames)
-
-		// Add tag groups and requests
-		for _, tagName := range tagNames {
-			requests := tags[tagName]
-
-			// Add tag group header (unless it's "untagged" and only one tag group)
-			if len(tagNames) > 1 || tagName != "untagged" {
-				tagDisplayName := "🏷️ " + tagName
-				if tagName == "untagged" {
-					tagDisplayName = "📄 untagged"
-				}
-				m.collections = append(m.collections, panels.CollectionItem{
-					Name:         "    " + tagDisplayName,
-					Type:         "tag",
-					FilePath:     "",
-					IsFolder:     false,
-					IsTagGroup:   true,
-					RequestIndex: -1,
-					IsExpanded:   false, // Collapsed by default
-					IsVisible:    false, // Hidden when parent folder is collapsed
-				})
-			}
-
-			// Sort requests by HTTP method priority (GET, POST, PUT, PATCH, DELETE, others)
-			sort.Slice(requests, func(i, j int) bool {
-				priorityI := getMethodPriority(requests[i].HTTP.Method)
-				priorityJ := getMethodPriority(requests[j].HTTP.Method)
-				if priorityI != priorityJ {
-					return priorityI < priorityJ
-				}
-				// If same method priority, sort alphabetically by name
-				return requests[i].Meta.Name < requests[j].Meta.Name
-			})
-
-			// Add requests under this tag
-			for _, request := range requests {
-				methodColor := getMethodColor(request.HTTP.Method)
-				indentLevel := "        "
-				if len(tagNames) == 1 && tagName == "untagged" {
-					indentLevel = "    " // Less indentation if no tag groups
-				}
-				// Calculate available width for collections panel (width/3 - 4 for padding)
-				availableWidth := m.width/3 - 4
-				displayName := formatRequestDisplayName(indentLevel, methodColor, request.Meta.Name, request.HTTP.Method, availableWidth)
-
-				m.collections = append(m.collections, panels.CollectionItem{
-					Name:         displayName,
-					Type:         "request",
-					FilePath:     requestPaths[request],
-					IsFolder:     false,
-					IsTagGroup:   false,
-					RequestIndex: requestIndexMap[request],
-					IsExpanded:   false, // Not applicable for requests
-					IsVisible:    false, // Hidden when parent tag/folder is collapsed
-				})
-			}
-		}
-	}
-
-	// Add root requests with tag grouping
-	if len(rootRequests) > 0 {
-		// Get sorted tag names
-		tagNames := make([]string, 0, len(rootRequests))
-		for tagName := range rootRequests {
-			tagNames = append(tagNames, tagName)
-		}
-		sort.Strings(tagNames)
-
-		// Add tag groups and requests
-		for _, tagName := range tagNames {
-			requests := rootRequests[tagName]
-			if len(requests) == 0 {
-				continue
-			}
-
-			// Add tag group header (unless it's "untagged" and only one tag group)
-			if len(tagNames) > 1 || tagName != "untagged" {
-				tagDisplayName := "🏷️ " + tagName
-				if tagName == "untagged" {
-					tagDisplayName = "📄 untagged"
-				}
-				m.collections = append(m.collections, panels.CollectionItem{
-					Name:         tagDisplayName,
-					Type:         "tag",
-					FilePath:     "",
-					IsFolder:     false,
-					IsTagGroup:   true,
-					RequestIndex: -1,
-					IsExpanded:   false, // Collapsed by default
-					IsVisible:    true,  // Root tags are visible
-				})
-			}
-
-			// Sort requests by HTTP method priority (GET, POST, PUT, PATCH, DELETE, others)
-			sort.Slice(requests, func(i, j int) bool {
-				priorityI := getMethodPriority(requests[i].HTTP.Method)
-				priorityJ := getMethodPriority(requests[j].HTTP.Method)
-				if priorityI != priorityJ {
-					return priorityI < priorityJ
-				}
-				// If same method priority, sort alphabetically by name
-				return requests[i].Meta.Name < requests[j].Meta.Name
-			})
-
-			// Add requests under this tag
-			for _, request := range requests {
-				methodColor := getMethodColor(request.HTTP.Method)
-				indentLevel := "    "
-				if len(tagNames) == 1 && tagName == "untagged" {
-					indentLevel = "  " // Less indentation if no tag groups
-				}
-				// Calculate available width for collections panel (width/3 - 4 for padding)
-				availableWidth := m.width/3 - 4
-				displayName := formatRequestDisplayName(indentLevel, methodColor, request.Meta.Name, request.HTTP.Method, availableWidth)
-
-				m.collections = append(m.collections, panels.CollectionItem{
-					Name:         displayName,
-					Type:         "request",
-					FilePath:     requestPaths[request],
-					IsFolder:     false,
-					IsTagGroup:   false,
-					RequestIndex: requestIndexMap[request],
-					IsExpanded:   false, // Not applicable for requests
-					IsVisible:    false, // Hidden when parent tag/folder is collapsed
-				})
-			}
-		}
-	}
+	data := LoadBruFiles(collectionsDir, m.width)
+	m.collections = data.Collections
+	m.bruRequests = data.BruRequests
 
 	if len(m.bruRequests) > 0 {
 		m.currentReq = m.bruRequests[0]
@@ -550,58 +282,6 @@ func (m *model) updateCollectionsViewport() {
 			m.collectionsViewport.SetYOffset(selectedVisibleIndex)
 		}
 	}
-}
-
-func getMethodColor(method string) string {
-	switch method {
-	case "GET":
-		return "🟢"
-	case "POST":
-		return "🟡"
-	case "PUT":
-		return "🔵"
-	case "DELETE":
-		return "🔴"
-	case "PATCH":
-		return "🟠"
-	default:
-		return "⚪"
-	}
-}
-
-func getMethodPriority(method string) int {
-	switch method {
-	case "GET":
-		return 1
-	case "POST":
-		return 2
-	case "PUT":
-		return 3
-	case "PATCH":
-		return 4
-	case "DELETE":
-		return 5
-	default:
-		return 6 // Other methods come last
-	}
-}
-
-func formatRequestDisplayName(indentLevel, methodEmoji, requestName, method string, availableWidth int) string {
-	// Create the left part: indentation + emoji + space + request name
-	leftPart := indentLevel + methodEmoji + " " + requestName
-	
-	// Create the right part: [METHOD]
-	rightPart := "[" + method + "]"
-	
-	// Calculate padding needed for right alignment
-	usedSpace := len(leftPart) + len(rightPart)
-	if usedSpace < availableWidth {
-		padding := availableWidth - usedSpace
-		return leftPart + strings.Repeat(" ", padding) + rightPart
-	}
-	
-	// If not enough space, just add one space
-	return leftPart + " " + rightPart
 }
 
 // Wrapper methods for filter operations
